requirements.txt
pandas
kagglehub
langchain
langchain-openai
chromadb
python-dotenv
tiktoken
streamlit